requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.1.4
yfinance==0.2.33
python-dotenv==1.0.0
schedule==1.2.0
gitpython==3.1.41
pyyaml==6.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
